backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx[http2]>=0.27.0
pydantic>=2.6.0
anthropic>=0.40.0
python-dotenv>=1.0.0
